Added filesize display in view files

+++ schema.sql
@@ -27,7 +27,7 @@ create table file (
   file_id integer primary key autoincrement,
   file_name text,
   file_path text,
-  file_size integer,
+  file_size numeric,
   file_today_date text
 );
 --Settings Data
